requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.0
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
pandas==2.1.4
ta==0.11.0
yfinance==0.2.33
ccxt>=4.0.0
APScheduler==3.10.4
